# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "VibeCoding社区 C 端（帖子、评论、点赞、通知、上传）与 B 端（内容管理、用户管理、数据统计）接口文档",
        "title": "VibeCoding 社区 API",
        "termsOfService": "https://github.com/vibecoding/community",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/vibecoding/community"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/admin/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B端-认证"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "username, password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_admin.AdminLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_admin.AdminLoginResp"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/posts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B端-内容"
                ],
                "summary": "管理后台帖子列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "20",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "normal|hidden|flagged",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/admin/posts/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B端-内容"
                ],
                "summary": "管理后台帖子详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vibecoding_community_internal_model.Post"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "B端-内容"
                ],
                "summary": "删除帖子",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B端-内容"
                ],
                "summary": "更新帖子状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_admin.UpdatePostStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B端-用户"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "C端-认证"
                ],
                "summary": "C端登录",
                "parameters": [
                    {
                        "description": "phone, code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_c.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_c.LoginResp"
                        }
                    }
                }
            }
        },
...
        "/api/v1/posts/{id}/comments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "C端-互动"
                ],
                "summary": "评论列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "C端-互动"
                ],
                "summary": "发表评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_c.CreateCommentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vibecoding_community_internal_model.Comment"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_vibecoding_community_internal_model.Comment": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "post_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_vibecoding_community_internal_model.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_vibecoding_community_internal_model.Post": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_vibecoding_community_internal_model.PostMedia"
                    }
                },
                "status": {
                    "$ref": "#/definitions/github_com_vibecoding_community_internal_model.PostStatus"
                },
                "type": {
                    "$ref": "#/definitions/github_com_vibecoding_community_internal_model.PostType"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_vibecoding_community_internal_model.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_vibecoding_community_internal_model.PostMedia": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "post_id": {
                    "type": "integer"
                },
                "sort_order": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_vibecoding_community_internal_model.PostStatus": {
            "type": "string",
            "enum": [
                "normal",
                "hidden",
                "flagged"
            ],
            "x-enum-varnames": [
                "PostStatusNormal",
                "PostStatusHidden",
                "PostStatusFlagged"
            ]
        },
        "github_com_vibecoding_community_internal_model.PostType": {
            "type": "string",
            "enum": [
                "text",
                "image",
                "video"
            ],
            "x-enum-varnames": [
                "PostTypeText",
                "PostTypeImage",
                "PostTypeVideo"
            ]
        },
        "github_com_vibecoding_community_internal_model.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_handler_admin.AdminLoginReq": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_handler_admin.AdminLoginResp": {
            "type": "object",
            "properties": {
                "admin": {},
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_handler_admin.UpdatePostStatusReq": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "normal, hidden, flagged",
                    "type": "string"
                }
            }
        },
        "internal_handler_c.CreateCommentReq": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler_c.LoginReq": {
            "type": "object",
            "required": [
                "code",
                "phone"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "internal_handler_c.LoginResp": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {}
            }
        }
    }
}